# public/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Reviews POC API",
        "version": "1.0"
    },
    "host": "reviews-poc.dev",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/reviews": {
            "get": {
                "tags": [
                    "review"
                ],
                "summary": "Display a listing of reviews.",
                "description": "Returns listing of reviews.",
                "operationId": "review.index",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of reviews per page",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "review"
                ],
                "summary": "Create new review",
                "description": "Creates a review.",
                "operationId": "review.create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Review to be created",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Review"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review created."
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reviews/{id}": {
            "get": {
                "tags": [
                    "review"
                ],
                "summary": "View a single review",
                "description": "Returns a single review.",
                "operationId": "review.view",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of review",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews."
                    },
                    "404": {
                        "description": "Review not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "review"
                ],
                "summary": "Update an existing review by ID",
                "description": "Updates a review.",
                "operationId": "review.update",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of review",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Review to be created",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Review"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review updated."
                    },
                    "404": {
                        "description": "Review not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "review"
                ],
                "summary": "Delete an existing review by ID.",
                "description": "Deletes a review.",
                "operationId": "review.delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of review",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review updated."
                    },
                    "404": {
                        "description": "Review not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "Review": {
            "required": [
                "name",
                "description"
            ],
            "properties": {
                "name": {
                    "description": "Review model",
                    "type": "string"
                },
                "description": {
                    "description": "Review model",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}